requirement.txt
flask==3.0.3
faiss-cpu==1.8.0
sentence-transformers==3.0.1
datasets==2.20.0
transformers==4.42.4
torch==2.3.1
numpy==1.26.4
huggingface-hub==0.23.4